requirements.txt
#Pi 3 explorer Radar
requests

#bluepy
git+https://github.com/IanHarvey/bluepy.git
pyserial